requirements.txt
# Train model requirements
tensorflow>=2.11.0
matplotlib
numpy
pandas
scikit-learn
seaborn


# fastAPI Requirements
fastapi
uvicorn[standard]
tensorflow
pillow
python-multipart
numpy